Worked on Calendar scene
Added particles for pages, wall and sprite for calendar object, camera and viewport, ability to see 2D and 3D in same scenes

Ran into an issue where the game cannot open due to "invalid access to key named 'load_scene'" or whatever. Super awesome. Gonna take a look at the tutorials of the guy who made the SceneManager+ plugin and see if I can make some sense out of it.